Integrated Apptentive Rate Dialog. Updated Crashlytics and Twitter libs. Add a comment button redesign. Hide user's own info in UsersProfile screen. ActionBarActivity -> AppCompatActivity. Bug fix: search button is moved in the overflow menu on smaller screens.

+++ app/src/main/java/com/apphunt/app/ui/fragments/CollectionsFragment.java
@@ -68,7 +68,7 @@ public class CollectionsFragment extends BaseFragment implements ViewPager.OnPag
             public void onTabSelected(TabLayout.Tab tab) {
                 int tag = (Integer) tab.getTag();
                 setTabSelectedIcon(tab);
-                pager.setCurrentItem((Integer) tag);
+                pager.setCurrentItem(tag);
                 resetTabIcon(previousSelectedTab);
                 previousSelectedTab = tab;
             }
